Refactoring provider to avoid IWorkContextAccessor
The IWorkContextAccessor interface works only when the call is made
in context of an HTTP request. While this is always true in case of
the web console, the standalone OrchardPs.exe application creates a
standalone AppDomain and calls into it without an HTTP request.
Therefore, the Orchard PS provider may not rely on this API.

Additionally, this changes fixes lookup of files in Orchard modules.
The dynamically created AppDomain loads module assemblies from the
Temporary Internet Files folder, so we need to tweak the logic which
determines whether the specified module is enabled.