Get features by id, not by name
This modifies the Get-OrchardFeature and Get-OrchardTheme cmdlets by
making them search features/themes by feature id by default, not by
name. Feature IDs are unique and they do not contain spaces, so they
are a lot easier to type at the console.

Additionally, this changeset adds an Id parameter to Get-OrchardFeature
and Get-OrchardTheme, so that users can explicitly specify to search by
Name or Id. Furthermore, the feature/theme ID is displayed in the output
of Get-OrchardFeature and Get-OrchardTheme.