Support for using PS session from code
This changeset adds better support for using the PowerShell engine
from code (instead of using OrchardPs.exe or the web console). It
introduces the BufferConsoleConnection class which implements the
IConsoleConnection interface using a set of memory buffers which may
be manipulated from code easily. Additionlly, it adds some extension
methods for IPsSessionManager to make it easy to create new PS sessions
from code.